Effective data tests moved to seperate class
